Crate ab_contracts_io_type

Modules

maybe_data

metadata

trivial_type

variable_bytes

variable_elements

Constants

MAX_ALIGNMENT

The maximum alignment supported by IoType types (16 bytes, corresponds to alignment of u128)

Traits

IoType

Trait that is used for types that are crossing host/guest boundary in contracts.

IoTypeOptional

Marker trait, companion to IoType that indicates the ability to store optional contents