ab_contracts_io_type
Enum IoTypeMetadata

#[repr(u8)]
pub enum IoTypeMetadata {

    Unit = 0,
    Bool = 1,
    U8 = 2,
    U16 = 3,
    U32 = 4,
    U64 = 5,
    U128 = 6,
    I8 = 7,
    I16 = 8,
    I32 = 9,
    I64 = 10,
    I128 = 11,
    Struct = 12,
    Struct1 = 13,
    Struct2 = 14,
    Struct3 = 15,
    Struct4 = 16,
    Struct5 = 17,
    Struct6 = 18,
    Struct7 = 19,
    Struct8 = 20,
    Struct9 = 21,
    Struct10 = 22,
    Struct11 = 23,
    Struct12 = 24,
    Struct13 = 25,
    Struct14 = 26,
    Struct15 = 27,
    Struct16 = 28,
    Enum = 29,
    Enum1 = 30,
    Enum2 = 31,
    Enum3 = 32,
    Enum4 = 33,
    Enum5 = 34,
    Enum6 = 35,
    Enum7 = 36,
    Enum8 = 37,
    Enum9 = 38,
    Enum10 = 39,
    Enum11 = 40,
    Enum12 = 41,
    Enum13 = 42,
    Enum14 = 43,
    Enum15 = 44,
    Enum16 = 45,
    Array8b = 46,
    Array16b = 47,
    Array32b = 48,
    ArrayU8x8 = 49,
    ArrayU8x16 = 50,
    ArrayU8x32 = 51,
    ArrayU8x64 = 52,
    ArrayU8x128 = 53,
    ArrayU8x256 = 54,
    ArrayU8x512 = 55,
    ArrayU8x1024 = 56,
    ArrayU8x2028 = 57,
    ArrayU8x4096 = 58,
    VariableBytes8b = 59,
    VariableBytes16b = 60,
    VariableBytes32b = 61,
    VariableBytes512 = 62,
    VariableBytes1024 = 63,
    VariableBytes2028 = 64,
    VariableBytes4096 = 65,
    VariableBytes8192 = 66,
    VariableBytes16384 = 67,
    VariableBytes32768 = 68,
    VariableBytes65536 = 69,
    VariableBytes131072 = 70,
    VariableBytes262144 = 71,
    VariableBytes524288 = 72,
    VariableBytes1048576 = 73,
    VariableBytes4194304 = 74,
    VariableBytes8388608 = 75,
    VariableBytes16777216 = 76,

}

Metadata types contained in TrivialType::METADATA and IoType::METADATA.

Metadata encoding consists of this enum variant treated as u8 followed by optional metadata encoding rules specific to metadata type variant (see variant’s description).

This metadata is sufficient to fully reconstruct hierarchy of the type in order to generate language bindings, auto-generate UI forms, etc.

Variants

Unit = 0

()

Bool = 1

bool

U8 = 2

u8

U16 = 3

u16

U32 = 4

u32

U64 = 5

u64

U128 = 6

u128

I8 = 7

i8

I16 = 8

i16

I32 = 9

i32

I64 = 10

i64

I128 = 11

i128

Struct = 12

struct

Structs are encoded af follows:

Length of struct name in bytes (u8)

Struct name as UTF-8 bytes

Number of fields (u8)

Each field is encoded follows:

Length of the field name in bytes (u8)

Field name as UTF-8 bytes

Recursive metadata of the field’s type

In tuple structs the first tuple element is treated as a field with name 0, second with 1 and so on, following above rules otherwise.

Compact encoding doesn’t encode names of structs and fields and their length.

Struct1 = 13

Similar to Self::Struct, but for exactly 1 struct fields and thus skips number of fields after struct name

Struct2 = 14

Similar to Self::Struct, but for exactly 2 struct fields and thus skips number of fields after struct name

Struct3 = 15

Similar to Self::Struct, but for exactly 3 struct fields and thus skips number of fields after struct name

Struct4 = 16

Similar to Self::Struct, but for exactly 4 struct fields and thus skips number of fields after struct name

Struct5 = 17

Similar to Self::Struct, but for exactly 5 struct fields and thus skips number of fields after struct name

Struct6 = 18

Similar to Self::Struct, but for exactly 6 struct fields and thus skips number of fields after struct name

Struct7 = 19

Similar to Self::Struct, but for exactly 7 struct fields and thus skips number of fields after struct name

Struct8 = 20

Similar to Self::Struct, but for exactly 8 struct fields and thus skips number of fields after struct name

Struct9 = 21

Similar to Self::Struct, but for exactly 9 struct fields and thus skips number of fields after struct name

Struct10 = 22

Similar to Self::Struct, but for exactly 10 struct fields and thus skips number of fields after struct name

Struct11 = 23

Similar to Self::Struct, but for exactly 11 struct fields and thus skips number of fields after struct name

Struct12 = 24

Similar to Self::Struct, but for exactly 12 struct fields and thus skips number of fields after struct name

Struct13 = 25

Similar to Self::Struct, but for exactly 13 struct fields and thus skips number of fields after struct name

Struct14 = 26

Similar to Self::Struct, but for exactly 14 struct fields and thus skips number of fields after struct name

Struct15 = 27

Similar to Self::Struct, but for exactly 15 struct fields and thus skips number of fields after struct name

Struct16 = 28

Similar to Self::Struct, but for exactly 16 struct fields and thus skips number of fields after struct name

Enum = 29

enum

Enums are encoded as follows:

Length of enum name in bytes (u8)

Enum name as UTF-8 bytes

Number of variants (u8)

Each enum variant as if it was a struct, see Self::Struct for details

Compact encoding doesn’t encode names of enums, variants and fields and their length.

Enum1 = 30

Similar to Self::Enum, but for exactly 1 enum variants and thus skips number of variants after enum name

Enum2 = 31

Similar to Self::Enum, but for exactly 2 enum variants and thus skips number of variants after enum name

Enum3 = 32

Similar to Self::Enum, but for exactly 3 enum variants and thus skips number of variants after enum name

Enum4 = 33

Similar to Self::Enum, but for exactly 4 enum variants and thus skips number of variants after enum name

Enum5 = 34

Similar to Self::Enum, but for exactly 5 enum variants and thus skips number of variants after enum name

Enum6 = 35

Similar to Self::Enum, but for exactly 6 enum variants and thus skips number of variants after enum name

Enum7 = 36

Similar to Self::Enum, but for exactly 7 enum variants and thus skips number of variants after enum name

Enum8 = 37

Similar to Self::Enum, but for exactly 8 enum variants and thus skips number of variants after enum name

Enum9 = 38

Similar to Self::Enum, but for exactly 9 enum variants and thus skips number of variants after enum name

Enum10 = 39

Similar to Self::Enum, but for exactly 10 enum variants and thus skips number of variants after enum name

Enum11 = 40

Similar to Self::Enum, but for exactly 11 enum variants and thus skips number of variants after enum name

Enum12 = 41

Similar to Self::Enum, but for exactly 12 enum variants and thus skips number of variants after enum name

Enum13 = 42

Similar to Self::Enum, but for exactly 13 enum variants and thus skips number of variants after enum name

Enum14 = 43

Similar to Self::Enum, but for exactly 14 enum variants and thus skips number of variants after enum name

Enum15 = 44

Similar to Self::Enum, but for exactly 15 enum variants and thus skips number of variants after enum name

Enum16 = 45

Similar to Self::Enum, but for exactly 16 enum variants and thus skips number of variants after enum name

Array8b = 46

Array [T; N] with up to 2^8 elements.

Arrays with up to 2^8 encoded as follows:

1 byte number of elements

Recursive metadata of contained type

Array16b = 47

Array [T; N] with up to 2^16 elements.

Arrays with up to 2^16 encoded as follows:

2 bytes number of elements (little-endian)

Recursive metadata of contained type

Array32b = 48

Array [T; N] with up to 2^32 elements.

Arrays with up to 2^32 encoded as follows:

4 bytes number of elements (little-endian)

Recursive metadata of contained type

ArrayU8x8 = 49

Compact alias for [u8; 8]

ArrayU8x16 = 50

Compact alias for [u8; 16]

ArrayU8x32 = 51

Compact alias for [u8; 32]

ArrayU8x64 = 52

Compact alias for [u8; 64]

ArrayU8x128 = 53

Compact alias for [u8; 128]

ArrayU8x256 = 54

Compact alias for [u8; 256]

ArrayU8x512 = 55

Compact alias for [u8; 512]

ArrayU8x1024 = 56

Compact alias for [u8; 1024]

ArrayU8x2028 = 57

Compact alias for [u8; 2028]

ArrayU8x4096 = 58

Compact alias for [u8; 4096]

VariableBytes8b = 59

Variable bytes with up to 2^8 bytes recommended allocation.

Variable bytes with up to 2^8 encoded as follows:

1 byte recommended allocation in bytes

VariableBytes16b = 60

Variable bytes with up to 2^16 bytes recommended allocation.

Variable bytes with up to 2^16 encoded as follows:

2 bytes recommended allocation in bytes (little-endian)

VariableBytes32b = 61

Variable bytes with up to 2^32 bytes recommended allocation.

Variable bytes with up to 2^8 encoded as follows:

4 bytes recommended allocation in bytes (little-endian)

VariableBytes512 = 62

Compact alias VariableBytes<512>

VariableBytes1024 = 63

Compact alias VariableBytes<1024>

VariableBytes2028 = 64

Compact alias VariableBytes<2028>

VariableBytes4096 = 65

Compact alias VariableBytes<4096>

VariableBytes8192 = 66

Compact alias VariableBytes<8192>

VariableBytes16384 = 67

Compact alias VariableBytes<16384>

VariableBytes32768 = 68

Compact alias VariableBytes<32768>

VariableBytes65536 = 69

Compact alias VariableBytes<65536>

VariableBytes131072 = 70

Compact alias VariableBytes<131072>

VariableBytes262144 = 71

Compact alias VariableBytes<262144>

VariableBytes524288 = 72

Compact alias VariableBytes<524288>

VariableBytes1048576 = 73

Compact alias VariableBytes<1048576>

VariableBytes4194304 = 74

Compact alias VariableBytes<4194304>

VariableBytes8388608 = 75

Compact alias VariableBytes<8388608>

VariableBytes16777216 = 76

Compact alias VariableBytes<16777216>

Implementations

impl IoTypeMetadata

pub const MAX_ALIGNMENT: usize = 4_096usize

This is helpful for host allocations that do not need to think about allocation alignment and can simply allocate all data structures at 4096 bytes alignment.

Without this types metadata would have to store alignment alongside other details, but since alignment beyond 4096 bytes is unlikely to be used in practice, metadata can be simplified.