ab_contracts_io_type::variable_bytes
Struct VariableBytes

pub struct VariableBytes<const RECOMMENDED_ALLOCATION: u32> { /* private fields */ }

Container for storing variable number of bytes.

RECOMMENDED_ALLOCATION is what is being used when host needs to allocate memory for call into guest, but guest may receive an allocation with more or less memory in practice depending on other circumstances, like when called from another contract with specific allocation specified.

Implementations

impl<const RECOMMENDED_ALLOCATION: u32> VariableBytes<RECOMMENDED_ALLOCATION>

pub fn size(&self) -> u32

pub fn capacity(&self) -> u32

pub fn get_initialized(&self) -> &[u8]

Try to get access to initialized bytes

pub fn get_initialized_mut(&mut self) -> &mut [u8]

Try to get exclusive access to initialized Data, returns None if not initialized

pub fn append(&mut self, bytes: &[u8]) -> bool

Append some bytes by using more of allocated, but currently unused bytes.

true is returned on success, but if there isn’t enough unused bytes left, false is.

pub fn truncate(&mut self, new_size: u32) -> bool

Truncate internal initialized bytes to this size.

Returns true on success or false if new_size is larger than Self::size().

pub fn as_mut_ptr(&mut self) -> &mut NonNull<u8>

Get exclusive access to underlying pointer with no checks.

Can be used for initialization with Self::assume_init() called afterward to confirm how many bytes are in use right now.

pub unsafe fn assume_init(&mut self, size: u32) -> Option<&mut [u8]>

Assume that the first size are initialized and can be read.

Returns Some(initialized_bytes) on success or None if size is larger than its capacity.

Safety

Caller must ensure size are actually initialized